moved diffy creation for demo into app.js and created root react component in index.js
