+++ demo/index.html
@@ -15,7 +15,7 @@
   </style>
 </head>
 <body>
-
+  <div id="app"></div>
   <canvas
     id="grid-canvas"
     width="1040"
