+++ package.json
@@ -28,7 +28,7 @@
   "homepage": "https://github.com/maniart/diffyjs#readme",
   "devDependencies": {
     "babel-cli": "^6.18.0",
-    "babel-core": "^6.18.0",
+    "babel-core": "^6.26.3",
     "babel-loader": "^6.2.5",
     "babel-polyfill": "^6.20.0",
     "babel-preset-es2015": "^6.18.0",
@@ -37,15 +37,20 @@
     "eslint": "^3.8.1",
     "eslint-config-airbnb": "^12.0.0",
     "eslint-loader": "^1.6.0",
-    "eslint-plugin-import": "^2.0.1",
+    "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jsx-a11y": "^2.2.3",
+    "eslint-plugin-react": "^7.11.1",
     "eslint-watch": "^2.1.14",
-    "express": "^4.14.0",
+    "express": "^4.16.3",
     "mocha": "*",
     "npm-run-all": "^3.1.2",
     "webpack": "^3.8.1",
     "webpack-dev-middleware": "1.6.1",
     "webpack-hot-middleware": "2.10.0",
     "worker-loader": "^0.7.1"
+  },
+  "dependencies": {
+    "react": "^16.5.2",
+    "react-dom": "^16.5.2"
   }
 }
